+++ wemake_python_styleguide/compat/functions.py
@@ -1,6 +1,7 @@
 import ast
-from typing import List, Union
+from typing import List, Tuple, Union
 
+from wemake_python_styleguide.compat.types import NodeWithTypeParams
 from wemake_python_styleguide.types import AnyAssignWithWalrus
 
 
@@ -11,3 +12,14 @@ def get_assign_targets(
     if isinstance(node, (ast.AnnAssign, ast.AugAssign, ast.NamedExpr)):
         return [node.target]
     return node.targets
+
+
+def get_type_param_names(  # pragma: py-lt-312
+    node: NodeWithTypeParams,
+) -> List[Tuple[ast.AST, str]]:
+    """Return list of type parameters' names."""
+    type_params = []
+    for type_param_node in getattr(node, 'type_params', []):
+        type_param_name = type_param_node.name
+        type_params.append((type_param_node, type_param_name))
+    return type_params