+++ wemake_python_styleguide/logic/tree/attributes.py
@@ -1,7 +1,8 @@
 import ast
 from typing import Iterable, Optional
 
-from wemake_python_styleguide.types import AnyChainable
+from wemake_python_styleguide.constants import SPECIAL_ARGUMENT_NAMES_WHITELIST
+from wemake_python_styleguide.types import AnyChainable, AnyVariableDef
 
 
 def _chained_item(iterator: ast.AST) -> Optional[ast.AST]:
@@ -33,3 +34,18 @@ def parts(node: AnyChainable) -> Iterable[ast.AST]:
         if chained_item is None:
             return
         iterator = chained_item
+
+
+def is_foreign_attribute(node: AnyVariableDef) -> bool:
+    """Tells whether this node is a foreign attribute."""
+    if not isinstance(node, ast.Attribute):
+        return False
+
+    if not isinstance(node.value, ast.Name):
+        return True
+
+    # This condition finds attributes like `point.x`,
+    # but, ignores all other cases like `self.x`.
+    # So, we change the strictness of this rule,
+    # based on the attribute source.
+    return node.value.id not in SPECIAL_ARGUMENT_NAMES_WHITELIST